feat: ability to defer playing till permission give; details below
workaround to browser bug, if not focus given to tab, aka deemed
inactive, the browser will not let you play media. use case is a player
deep links to your scene and switches windows while loading. the video
will try to play, the browser will stop it, and now the scene is out of
sync as it thinks playing==true but its not.